Prevent auto-save race condition by deferring editor creation until provider sync
The BlockNote editor was being created and bound to an empty Y.Doc before
the Hocuspocus provider synced with the server. This created a possible race condition
where auto-save could trigger and send empty updates to the server before
receiving the initial document state, causing data loss.

Split OpBlockNoteContainer into two components using React's
component mounting lifecycle as a gate:

- OpBlockNoteContainer: Pure sync state management. Shows skeleton loader
  while waiting for provider.on('synced'). Only renders BlockNoteEditor
  subcomponent after sync completes.

- BlockNoteEditor: Extracted editor logic with all BlockNote hooks. Only
  mounts when parent confirms sync is done, ensuring useCreateBlockNote
  is called with a Y.Doc that has already received server state.

This leverages React's component lifecycle guarantee: child components only
mount when parent renders them, preventing the hook call until sync is
guaranteed complete.

https://community.openproject.org/wp/69724
